+++ packages/insomnia/src/ui/routes/$organizationId.project.$projectId.git-actions.tsx
@@ -1,7 +1,6 @@
 import { fromUrl } from 'hosted-git-info';
 import { type ActionFunction, type LoaderFunction, redirect } from 'react-router';
 
-import { gitCredentials } from '../../models';
 import type { GitRepository } from '../../models/git-repository';
 import { WorkspaceScopeKeys } from '../../models/workspace';
 import { type GitLogEntry } from '../../sync/git/git-vcs';
@@ -467,66 +466,4 @@ export const diffFileLoader: LoaderFunction = async ({ request, params }): Promi
   const staged = urlParams.get('staged') === 'true';
 
   return window.main.git.diffFileLoader({ projectId, workspaceId, filepath, staged });
-};
-
-export const loadGitHubCredentials: LoaderFunction = async () => {
-  const credentials = await gitCredentials.getByProvider('github');
-
-  return credentials;
-};
-
-export const initSignInToGitHub: ActionFunction = async () => {
-  await window.main.git.initSignInToGitHub();
-
-  return null;
-};
-
-export const completeSignInToGitHub: ActionFunction = async ({ request }) => {
-  const { code, state } = (await request.json()) as { code: string; state: string; path: string };
-  await window.main.git.completeSignInToGitHub({
-    code,
-    state,
-  });
-
-  return null;
-};
-
-export const signOutOfGitHub: ActionFunction = async () => {
-  await window.main.git.signOutOfGitHub();
-
-  return null;
-};
-
-export const loadGitLabCredentials: LoaderFunction = async () => {
-  const credentials = await gitCredentials.getByProvider('gitlab');
-
-  return credentials;
-};
-
-export const initSignInToGitLab: ActionFunction = async () => {
-  await window.main.git.initSignInToGitLab();
-
-  return null;
-};
-
-export const completeSignInToGitLab: ActionFunction = async ({ request }) => {
-  const { code, state } = (await request.json()) as { code: string; state: string; path: string };
-  await window.main.git.completeSignInToGitLab({
-    code,
-    state,
-  });
-
-  return null;
-};
-
-export const signOutOfGitLab: ActionFunction = async () => {
-  await window.main.git.signOutOfGitLab();
-
-  return null;
-};
-
-export const loadGitCredentials: LoaderFunction = async () => {
-  const credentials = await gitCredentials.all();
-
-  return credentials;
 };
